refactor(load-tester): 디렉토리 리팩토링
