+++ load-tester/src/main/java/org/prebird/loadtester/controller/LoadTestWebController.java
@@ -1,8 +1,9 @@
-package org.prebird.loadtester;
+package org.prebird.loadtester.controller;
 
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.prebird.loadtester.domain.LoadTestRepository;
+import org.prebird.loadtester.dto.LoadTestDto;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
