+++ load-tester/src/main/java/org/prebird/loadtester/dto/LoadTestDto.java
@@ -1,4 +1,4 @@
-package org.prebird.loadtester;
+package org.prebird.loadtester.dto;
 
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
